Modules/Alpha_Detector/requirements.txt
# Alpha Detector Module Dependencies
# Phase 1: Foundation

# Database
psycopg2-binary==2.9.10

# WebSocket
websockets==15.0.1

# Async support
asyncio==4.0.0

# Environment management
python-dotenv==1.1.1

# Data processing
pandas==2.2.0
numpy==1.26.4

# Configuration
PyYAML==6.0.2

# Logging
structlog==24.1.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.5
